web_app/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.35
instagrapi==2.2.1
apscheduler==3.10.4
requests==2.32.4
python-dotenv==1.0.0
Pillow==10.0.1
opencv-python==4.8.1.78
cryptography==43.0.3
jinja2==3.1.4
aiofiles==23.2.1
pydantic==2.11.7
pyotp==2.9.0
pymysql==1.1.1
